+++ DNSRdataTypeCNAME.java
@@ -21,7 +21,7 @@ public class DNSRdataTypeCNAME extends DNSRdata
 
     @Override
     /** implement the abstract method. */
-    public int getDataLength()
+    public short getDataLength()
     {
         //for research project, no need for implementation here.
         return 0;
