+++ DNSRdataTypeMINFO.java
@@ -28,7 +28,7 @@ public class DNSRdataTypeMINFO extends DNSRdata
 
     @Override
     /** implement the abstract method. */
-    public int getDataLength()
+    public short getDataLength()
     {
         //for research project, no need for implementation here.
         return 0;
