+++ DNSRdataTypeNullSkippable.java
@@ -30,7 +30,7 @@ public class DNSRdataTypeNullSkippable extends DNSRdata
 
     @Override
     /** implement the abstract method. */
-    public int getDataLength()
+    public short getDataLength()
     {
         //for research project, no need for implementation here.
         return 0;
